Use uint32 for frame ID in F1 packet header
An unsigned 32-bit integer for the frame identifier in F1 packet headers
was incorrectly assumed to be an 8-bit unsigned integer. This previously
required a need to throw away 3 bytes before examining the packet body.
Examining frequent overflows in the frame identifier and incorrect
values for the player car index highlighted this issue.